+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.9.60
 
 * Spell Effect: Create Bonfire effect triggered at start and end, and save did not happen for damage on entry.
+* Effect Bardic Inspiration: did not work for skill bonuses.
 
 # 2.9.59