+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.9.61
+
+* Feature Effects: Blessed Strikes (Cleric).
+
 # 2.9.60
 
 * Spell Effect: Create Bonfire effect triggered at start and end, and save did not happen for damage on entry.